debugging

+++ model.py
@@ -291,7 +291,7 @@ class DeepVO(object):
     def categorical_accuracy(self, ys, y_pred):
         y_true = ys[0]
         print('here', y_true, y_pred)
-        y_pred = tf.argmax(y_pred, -1)
+        y_pred = tf.to_int64(tf.argmax(y_pred, -1))
         same_cat = tf.equal(y_true, y_pred)
         return tf.reduce_mean(tf.to_float(same_cat))
